+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Added
 
 - environmentvariablesprefixedby option
+- keep_git_dir option
 
 ## [0.2.0] - 2022-11-01